Move `step` under `pipeline`
